[msbuild] Verify Mac Catalyst apps like we verify macOS apps. Fixes #17829.
* For Mac Catalyst apps we must verify the code signature like we do it for
  macOS apps. Fixes #17829.
* Enable nullability and fix code accordingly.
* Merge the CodesignVerify and CodesignVerifyTaskBase classes to simplify the
  code.

Submission test is added as well:

* https://github.com/xamarin/maccore/pull/2656
* https://github.com/xamarin/SubmissionSamples/pull/51

Fixes https://github.com/xamarin/xamarin-macios/issues/17829.
